Issue #12: AWS autoscaling scripts refactoring
