+++ scripts/autoscaling/aws/node_reassign.py
@@ -12,87 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import boto3
 import argparse
-import pykube
 
-RUN_ID_LABEL = 'runid'
-AWS_REGION_LABEL = 'cloud_region'
-
-
-def find_and_tag_instance(ec2, old_id, new_id):
-    response = ec2.describe_instances(Filters=[{'Name': 'tag:Name', 'Values': [old_id]},
-                                               {'Name': 'instance-state-name', 'Values': ['pending', 'running']}])
-    tags = [{'Key': 'Name', 'Value': new_id}]
-    if len(response['Reservations']) > 0:
-        ins_id = response['Reservations'][0]['Instances'][0]['InstanceId']
-        ec2.create_tags(
-            Resources=[ins_id],
-            Tags=tags
-        )
-        return ins_id
-    else:
-        raise RuntimeError("Failed to find instance {}".format(old_id))
-
-
-def verify_regnode(kube_api, ec2, ins_id):
-    response = ec2.describe_instances(InstanceIds=[ins_id])
-    nodename = response['Reservations'][0]['Instances'][0]['PrivateDnsName'].split('.', 1)[0]
-    nodename_full = response['Reservations'][0]['Instances'][0]['PrivateDnsName']
-
-    exist_node = False
-    ret_namenode = ""
-    node = pykube.Node.objects(kube_api).filter(field_selector={'metadata.name': nodename})
-    if len(node.response['items']) > 0:
-        exist_node = True
-        ret_namenode = nodename
-    node_full = pykube.Node.objects(kube_api).filter(field_selector={'metadata.name': nodename_full})
-    if len(node_full.response['items']) > 0:
-        exist_node = True
-        ret_namenode = nodename_full
-    if not exist_node:
-        raise RuntimeError("Failed to find Node {}".format(ins_id))
-    return ret_namenode
-
-
-def find_instance_name(ec2, ins_id):
-    response = ec2.describe_instances(InstanceIds=[ins_id])
-    nodename = response['Reservations'][0]['Instances'][0]['PrivateDnsName'].split('.', 1)[0]
-    return nodename
-
-
-def change_label(api, nodename, new_id, aws_region):
-    obj = {
-        "apiVersion": "v1",
-        "kind": "Node",
-        "metadata": {
-            "name": nodename,
-            "labels": {
-                RUN_ID_LABEL: new_id
-            }
-        }
-    }
-    node = pykube.Node(api, obj)
-    node.labels[RUN_ID_LABEL] = new_id
-    node.labels[AWS_REGION_LABEL] = aws_region
-    node.update()
-
-
-def get_aws_region(api, run_id):
-    nodes = pykube.Node.objects(api).filter(selector={RUN_ID_LABEL: run_id})
-    if len(nodes.response['items']) == 0:
-        raise RuntimeError('Cannot find node matching RUN ID %s' % run_id)
-    node = nodes.response['items'][0]
-    labels = node['metadata']['labels']
-    if AWS_REGION_LABEL not in labels:
-        raise RuntimeError('Node %s is not labeled with AWS Region' % node['metadata']['name'])
-    return labels[AWS_REGION_LABEL]
-
-
-def get_kube_api():
-    api = pykube.HTTPClient(pykube.KubeConfig.from_file("~/.kube/config"))
-    api.session.verify = False
-    return api
+from pipeline.autoscaling import awsprovider, kubeprovider, utils
 
 
 def main():
@@ -103,12 +25,16 @@ def main():
     old_id = args.old_id
     new_id = args.new_id
 
-    kube_api = get_kube_api()
-    aws_region = get_aws_region(kube_api, old_id)
-    ec2 = boto3.client('ec2', region_name=aws_region)
-    ins_id = find_and_tag_instance(ec2, old_id, new_id)
-    nodename = verify_regnode(kube_api, ec2, ins_id)
-    change_label(kube_api, nodename, new_id, aws_region)
+    kube_provider = kubeprovider.KubeProvider()
+
+    cloud_region = kube_provider.get_cloud_region_by_node_name(args.node_name)
+    cloud_provider = awsprovider.AWSCloudProvider(cloud_region)
+
+    ins_id = cloud_provider.find_and_tag_instance(old_id, new_id)
+    nodename, nodename_full = cloud_provider.get_instance_names(ins_id)
+
+    nodename = kube_provider.verify_regnode(ins_id, nodename, nodename_full)
+    kube_provider.change_label(nodename, new_id, cloud_region)
 
 
 if __name__ == '__main__':
