+++ scripts/autoscaling/azure/terminate_node.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import pykube
 import argparse
 from pipeline.autoscaling import azureprovider, kubeprovider
